pyproject.toml
[project]
name = "example-1"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "requests (>=2.32.5,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
lint = [
    "flake8 (>=7.3.0,<8.0.0)",
    "black (>=25.12.0,<26.0.0)",
    "isort (>=7.0.0,<8.0.0)",
    "mypy (>=1.19.1,<2.0.0)"
]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)"
]

[tool.black]
line-length = 119
exclude = ".git"

[tool.isort]
line_length = 119

[tool.mypy]
disallow_untyped_defs = true
warn_return_any = true
exclude = "venv"
